--- ui/src/components/Payloads/NewPayloadForm.css
@@ -1,94 +1,0 @@
-.payload-form-container {
-    margin-top: calc(var(--header-height) + 1rem);
-    display: flex;
-    flex-direction: column;
-    padding: 1rem;
-    max-width: 980px;
-    margin-left: auto;
-    margin-right: auto;
-    background-color: var(--surface-1);
-    color: var(--text-color);
-    border-radius: 12px;
-    border: 1px solid var(--border-subtle);
-    box-shadow: 0 6px 16px rgba(0, 0, 0, 0.08);
-    max-height: calc(100vh - var(--header-height) - 3rem);
-    overflow-y: auto;
-}
-
-.payload-form-container input[type="text"],
-.payload-form-container textarea,
-.payload-form-container select {
-    width: 100%;
-    padding: 10px;
-    margin-bottom: 1rem;
-    border: 1px solid var(--border-subtle);
-    border-radius: 8px;
-    box-sizing: border-box;
-    background-color: var(--surface-2);
-    color: var(--text-color);
-    transition: border-color 0.3s ease, background-color 0.3s ease, color 0.3s ease;
-}
-
-.payload-form-container input[type="text"]::placeholder,
-.payload-form-container textarea::placeholder {
-    color: var(--placeholder-color);
-    opacity: 0.7;
-}
-
-.payload-form-container button {
-    padding: 0.6rem 1rem;
-    font-size: 0.95rem;
-    border: none;
-    border-radius: 8px;
-    background-color: var(--accent);
-    color: #fff;
-    cursor: pointer;
-    transition: background-color 0.3s ease, color 0.3s ease;
-}
-
-.payload-form-container button:disabled {
-    background-color: #cccccc;
-    color: #666666;
-    cursor: not-allowed;
-}
-
-.payload-form-container button:hover:not(:disabled) {
-    background-color: var(--accent-strong);
-}
-
-.notification {
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    color: white;
-    padding: 10px;
-    border-radius: 5px;
-    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
-    z-index: 9999;
-}
-
-.notification.success {
-    background-color: var(--success-color);
-}
-
-.notification.error {
-    background-color: var(--error-color);
-}
-
-.loading-indicator {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 1.2em;
-    color: var(--text-muted);
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    background: var(--surface-1);
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.2);
-    z-index: 1000;
-}
